requirements.txt
cryptography==3.3.1
Django==3.1.4
django-cors-headers==3.6.0
django-crontab==0.7.1
django-rest-framework==0.1.0
django-storages==1.11.1
djangorestframework==3.12.2
drf-yasg==1.20.0
html5lib==1.1
hyperlink==21.0.0
idna==2.10
imageio==2.9.0
imageio-ffmpeg==0.4.3
incremental==17.5.0
inflection==0.5.1
Jinja2==2.11.2
Pillow==8.1.0
python-dateutil==2.8.1
python-http-client==3.3.1
service-identity==18.1.0
sqlparse==0.4.1
